fix crashing bug in win32unix/select.c
git-svn-id: http://caml.inria.fr/svn/ocaml/version/4.01@14374 f963ae5c-01c2-4b8c-9fe0-0dff7051ff02

+++ otherlibs/win32unix/select.c
@@ -602,7 +602,7 @@ LPSELECTDATA socket_poll_add (LPSELECTDATA lpSelectData,
 {
   LPSELECTDATA res;
   LPSELECTDATA candidate;
-  DWORD i;
+  long i;
   LPSELECTQUERY aQueries;
 
   res = lpSelectData;
